+++ src/common/Modes.hpp
@@ -12,6 +12,8 @@ public:
     bool isNightly{};
     bool isPortable{};
 
+    bool useChatterinoDirectory{};
+
     /// Marked by the line `externally-packaged`
     ///
     /// The externally packaged mode comes with the following changes: