+++ src/singletons/Paths.cpp
@@ -105,6 +105,15 @@ void Paths::initRootDirectory()
                                      path.toStdString() + "\"");
         }
 
+        if (Modes::instance().useChatterinoDirectory)
+        {
+#ifdef Q_OS_WIN
+            path.replace("bajerino", "Chatterino");
+#else
+            path.replace("bajerino", "chatterino");
+#endif
+        }
+
 // create directory Chatterino2 instead of Chatterino on windows because the
 // ladder one is takes by Chatterino 1 already
 #ifdef Q_OS_WIN